% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{add.edges}
\alias{add.edges}
\alias{add_edges}
\title{add edges}
\usage{
add.edges(object, edges, ...)
}
\arguments{
\item{object}{igraph/mgraph object}

\item{edges}{various data type
\itemize{
\item sequence of integer or character: like igraph add.edges
\item edge names: each name is a pair of nodes seperated by "|" such as "A|B"
\item data.frame or matrix: first 2 columns are treated as sources and targets.
}}
}
\value{
graph object
}
\description{
add edges for igraph or mgraph object
}
\details{
Versatile add edges fucntion: accept sequences (like igraph::add.edges), vertex names or data.frame (first 2 columns are \code{from} and \code{to}). Refer to \code{igraph::add.edges} for other details.
}
\author{
ZG Zhao
}
