% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{add.reactions}
\alias{add.reactions}
\title{append reaction}
\usage{
add.reactions(object, ds, t, genes, name = NULL, reversible = FALSE)
}
\arguments{
\item{object}{ReactionList or equivalent list}

\item{ds}{character (substrate name) or data.frame with at least these columns: from, to, genes, reversible}

\item{t}{character, product name (ignore if ds is a data.frame)}

\item{genes}{character vector, names of genes involved in the reaction  (ignore if ds is a data.frame)}

\item{name}{additional identifer(s) for reaction  (ignore if ds is a data.frame)}

\item{reversible}{TRUE/FALSE (default). If TRUE, add reactions in both directions. Ignore if ds is a data.frame.}
}
\value{
ReactionList object
}
\description{
Append reaction to reaction lists
}
\details{
Useful for adding to network spontaneous reactions without associated genes.
}
\author{
ZG Zhao
}
