% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_cut_genes}
\alias{all_cut_genes}
\title{list all gene cut sets}
\usage{
all_cut_genes(object)
}
\arguments{
\item{object}{xgraph object}
}
\value{
list
\itemize{
\item min.number
\item min.sets
\item all.sets
}
}
\description{
This function list all gene cut sets. Remove any set of genes will disconnet primary substrates and end products.
}
\details{
Set substrates and products before using this function.
}
\author{
ZG Zhao
}
