% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_spaths_list}
\alias{all_spaths_list}
\alias{all_spaths_nodes}
\alias{all_spaths_edges}
\title{info of all simple paths}
\usage{
all_spath_list(obj, from, to, mc.cores)
all_spath_nodes(obj, from, to, mc.cores)
all_spath_nodes(obj)
all_spath_edges(obj, from, to, mc.cores)
all_spath_edges(obj)
}
\arguments{
\item{obj}{igraph/mgraph object or list returned by all_spaths_list}

\item{from}{character vector of length 1, name of start node}

\item{to}{character vector of length 1, name of end node}

\item{mc.cores}{refer to \code{\link{mclapply}}}
}
\value{
character list or vector
}
\description{
simple paths info: list, nodes and edges
}
\details{
Friendly wrapers for \code{\link{igraph::all_simple_paths}}.
}
\examples{
library(gmetab)
d.path <- file.path(path.package("gmetab"), "KEGG")
gg <- mgraph_from_kos("ko00010", d.path)
vv <- vnames(gg)
(splist <- all_spaths_list(gg, vv[1], vv[6]))
all_spaths_nodes(splist)
all_spaths_nodes(gg, vv[1], vv[6])
all_spaths_edges(splist)
all_spaths_edges(gg, vv[1], vv[6])
}
\author{
ZG Zhao
}
