% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_stcuts}
\alias{all_stcuts}
\title{list all gene cut sets}
\usage{
all_stcuts(object, s, t)
}
\arguments{
\item{object}{xgraph object}

\item{s}{source node}

\item{t}{target node}
}
\value{
list
\itemize{
\item min.genes
\item min.edges
\item edge.cuts
\item gene.cuts
}
}
\description{
Get st_cuts info: edges and genes
}
\details{
This is a wrapper function of \code{st_cuts} define in igraph package.
}
\author{
ZG Zhao
}
