% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{as_mreacts}
\alias{as_mreacts}
\title{generate MReactions object}
\usage{
as_mreacts(kinfo, org = "ko")
}
\arguments{
\item{kinfo}{KEGGmeta or list object. If feeding a list object, make sure each list elements are also lists containing substrate, product, gene and reversible elements.}

\item{org}{character, a KEGG organism string. Use if only kinfo is a plain list.}
}
\value{
MReactions object
}
\description{
Mostly for internal use.
}
\details{
"MReactions" (mreacts) is S4 class designed for mgraph. Reactions are presented in different forms from those of \code{\link{KEGGmeta}}.
}
\author{
ZG Zhao
}
