% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_classes.R
\name{class_KEGGmeta}
\alias{class_KEGGmeta}
\alias{KEGGmeta}
\title{S4 class: KEGGmeta}
\description{
"KEGGmeta" is S4 class designed for holding KEGG pathway information: entries, reactions, relations, graphics and other general information (pathInfo).
}
\details{
Typically, a KEGGmeta object can be obtained by \code{\link{kmeta_from_ko}} function, which will download and parse the KEGG xml file.
}
\seealso{
\code{\link{KEGGdata}}, \code{\link{ReactionSet}}
}
\author{
zhao
}
