% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_classes.R
\name{class_keggPATH}
\alias{class_keggPATH}
\alias{keggPATH}
\title{S4 class: keggPATH}
\description{
"keggPATH" is S4 class designed for holding KEGG pathway information: entries, reactions, relations, graphics and other general information (pathInfo).
}
\details{
Typically, a keggPATH object can be obtained by \code{\link{make_mpath}} function, which will download and parse the KEGG xml file.
}
\seealso{
\code{\link{KEntityList}}, \code{\link{ReactionSet}}
}
\author{
zhao
}
