% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_classes.R
\name{class_ReactionSet}
\alias{class_ReactionSet}
\title{S4 class: ReactionSet}
\description{
ReactionSet: S4 class holding reaction infos.
}
\details{
Three slots: reaction, organism and alias.
\itemize{
\item reaction: list of reaction info (substrate, product and genes)
\item organism: KEGG organism identifier
\item alias: named list of chemical aliases. Set when merging chemicals.
}
}
\author{
ZG Zhao
}
