% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{delete.edges}
\alias{delete.edges}
\alias{delete_edges}
\title{delete edges}
\usage{
delete.edges(object, es)
}
\arguments{
\item{object}{igraph/mgraph object}

\item{es}{vector: edge ids (integer) or names (character)}
}
\value{
igraph/mgraph object
}
\description{
delete edges from igraph or mgraph object
}
\details{
Same as functions in igraph package exception for retaining graph attributes. Refer to \code{?igraph::delete.edges}
}
\author{
ZG Zhao
}
