% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{edata}
\alias{edata}
\title{edge data}
\usage{
edata(g, a.name, e.names)
}
\arguments{
\item{g}{igraph/mgraph object}

\item{a.name}{character, attribute name}

\item{e.names}{vector of character (edge names) or integer (indices)}
}
\description{
Get or set edge attribute/data
}
\details{
Friendly function for edge data/attributes retrieving or setting. See "Example".
}
\examples{
library(gmetab)
d.path <- file.path(path.package("gmetab"), "KEGG")
gg <- make_mgraph("ko00010", d.path)
## igraph style
E(gg)$reaction
E(gg)$reaction[1:3]
E(gg)$MP <- 0
## gmetab style
edata(gg, "reaction")
edata(gg, "reaction", 1:3)
(ee <- enames(gg))
edata(gg, "reaction", ee[1:3])
edata(gg, "MP") <- 1
edata(gg, "MP")
}
\seealso{
\code{\link{vdata}}
}
\author{
ZG Zhao
}
