% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{enames}
\alias{enames}
\title{edge names}
\usage{
enames(object)
}
\arguments{
\item{object}{mgraph object}
}
\description{
get edge names
}
\details{
Similar functions: \code{\link{vnames}}, \code{\link{enames}}, \code{\link{rnames}}, \code{\link{vcount}}, \code{\link{ecount}}
}
\author{
ZG Zhao
}
