% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.struct.R
\name{FP_path}
\alias{FP_path}
\title{Calculate FP of a metabolic pathway}
\usage{
FP_path(
  g,
  substrates,
  products,
  simplify = TRUE,
  algorithm = "default",
  PPsubs = 1,
  silent = FALSE
)
}
\arguments{
\item{g}{graphMET object}

\item{substrates}{character vector. Node names representing substrates or starting points of the pathway.}

\item{products}{character vector. Node names representing products or end points of the pathway.}

\item{simplify}{TRUE/FALSE. If TRUE (default), simplify network before PGM construction and FP calculation.}

\item{algorithm}{character-string, one of "default", "forward" and "PGM"}

\item{PPsubs}{numeric vector, inital PP for substrates.}
}
\value{

}
\description{
FP (Failure Probability) is the probality of a pathway unable to get final products from given substrates. FP is calculated by probability propagation different algorithms. If more than one substrates (product) are given, they are treated equally and merged into a single "virtual substrate (product)"; all redundant nodes and edges will be removed accordingly. Please refer to \code{\link{mgraph_set_chems}} for more detail about merging of chemicals.
}
\details{
When error occurs, an integer (error code) will be returned, allowing smooth running of the function in parallel computing. The codes are:
\itemize{
\item 100: non graphMET object was feed or generated
\item 200: no node found
\item 300: no edge found
\item 400: substrates or products not presented in the pathway
\item 500: substrates and products not connected at all
}
}
\author{
ZG Zhao
}
