% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helpers_general}
\alias{helpers_general}
\alias{Organism}
\alias{Species}
\alias{Compounds}
\alias{Chemicals}
\alias{Reactions}
\alias{Genes}
\title{general helper function}
\description{
general helpers for retrieving data from keggPATH, reactions list or mgraph object
}
\details{
All functions below are generic functions. Function names are self-explanatory:
\itemize{
\item pathInfo(object): pathway info
\item Organism(object): organism name
\item Species(object): alias of Organism
\item Genes(object): all genes/orthologs
\item Compounds(object): all compounds
\item Chemicals(object): alias of Compounds
\item Reactions(object, list.only=TRUE): reactions list or ReactionSet object (list.only=FALSE for mgraph object)
}
}
\examples{
library(gmetab)
d.path <- file.path(path.package("gmetab"), "KEGG")
pp <- make_mpath("ko00010", d.path)
##
## KEGG pathway data
pathInfo(pp)
Organism(pp)
Compounds(pp)
Chemicals(pp)
Genes(pp)
Reactions(pp)
##
## metabolic graph
gg <- make_mgraph(pp)
Organism(gg)
Compounds(gg)
Chemicals(gg)
Genes(gg)
Reactions(gg)
Reactions(gg, list.only=FALSE)
}
\author{
ZG Zhao
}
