% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{helpers_is}
\alias{helpers_is}
\alias{is.klist}
\alias{is.rset}
\alias{is.igraph}
\alias{is.mgraph}
\alias{is.xgraph}
\alias{is.chemset}
\title{object test}
\value{
TRUE/FALSE
}
\description{
helper functions for object test
}
\details{
Most function names are self-explanatory:
\itemize{
\item is.mpath: keggPATH test
\item is.klist: KEGGtest test
\item is.rset: ReactionSet test
\item is.igraph(x)
\item is.mgraph(x)
\item is.xgraph: igraph or mgraph
\item is.chemset: test if primary substrates and end products were set.
\item is.empty: test whether all objects are "empty". "Empty" objects are those of NULL, all NA, zero length and string containing spaces only. Graph objects without any edge are also regarded as "empty". S4 objects are not "empty" whatever data they have.
}
}
\author{
ZG Zhao
}
