% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empties.R
\name{is.empty}
\alias{is.empty}
\title{test empty}
\usage{
is.empty(object, ...)
}
\arguments{
\item{object}{vector, matrix, data.frame, list, ReactionSet, graph ...}

\item{...}{multiple objects}
}
\value{
TRUE/FALSE
}
\description{
Uniform function for empty object test
}
\details{
"Empty" refers to zero items or empty strings if character vector is given.
}
\author{
ZG Zhao
}
