% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGG_get}
\alias{KEGG_get}
\title{KEGG download API}
\usage{
KEGG_get(
  qid,
  d.path = "KEGG",
  f.type = c("kgml", "image", "htext"),
  force.download = FALSE
)
}
\arguments{
\item{qid}{KEGG query id such as KEGG ortholog id like "ko00010", gene name like "ath:AT1G00010" or organism name like "ath" (for KEGG brite file only).}

\item{d.path}{Destination path (folder) to save or load downloaded file.}

\item{f.type}{Character string, file type of KEGG entry: "kgml", "aaseq", "ntseq" or "image".}

\item{force.download}{TRUE/FALSE (default). Force update if TRUE and file exists.}
}
\value{
path to the saved file.
}
\description{
API for downloding KEGG info with KEGG offical rest API.
}
\details{
Download XML, image (png), AA sequence or NT sequence file from KEGG website.
An alternative function is keggGet implemented in "KEGGREST" package.
}
\author{
ZG Zhao
}
