% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGG_ontology}
\alias{KEGG_ontology}
\title{KEGG ontology as data.frame or list}
\arguments{
\item{d.path}{see \code{\link{KEGG_get}}}

\item{force.download}{see \code{\link{KEGG_get}}}
}
\description{
Functions
\itemize{
\item konto_table(d.path, force.download)
\item konto_list(d.path, force.download)
}
}
\details{
Download (if file not exists) and parse ko00001.keg
}
\author{
ZG Zhao
}
