% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeta.R
\name{kmeta_from_ko}
\alias{kmeta_from_ko}
\title{Generate KEGGmeta object}
\usage{
kmeta_from_ko(ko, d.path = "KEGG")
}
\arguments{
\item{ko}{character, KEGG pathway ortholog (ko) index string such as ko00010, 00010 or ath00010}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
KEGGmeta object
}
\description{
Refer to \code{\link{KEGGmeta}} for class information.
}
\details{
This function reads from local folder (d.path) a KEGG xml file, which will be downloaded if not exists. Make sure internet is accessible.
}
\author{
ZG Zhao
}
