% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_classes.R
\name{kmeta_x_org}
\alias{kmeta_x_org}
\title{General to organism-specific KEGGmeta conversion}
\usage{
kmeta_x_org(kinfo, org = NA, d.path = "KEGG")
}
\arguments{
\item{kinfo}{a KEGGmeta object}

\item{org}{abbreviation of an organism such as "ath"}

\item{d.path}{file path for \code{\link{KEGG_get}}}
}
\value{
KEGGmeta object
}
\description{
Convert KOG names of a KEGGmeta object to real gene names
}
\details{
In KEGGmeta objects, genes involved in reactions are generally represented by KEGG entry names (k genes or KOG).
}
\author{
ZG Zhao
}
