% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{kogs_list}
\alias{kogs_list}
\title{KOG to gene mapping list for an organism}
\usage{
kogs_list(org, d.path = "KEGG", KOGs = NA)
}
\arguments{
\item{org}{character, KEGG organism abbreviation}

\item{d.path}{file path for \code{\link{KEGG_get}}}

\item{KOGs}{character vector, names of KOGs. Subset if not empty.}
}
\value{
list of genes named by KOGs.
}
\description{
Get KOG to gene mapping list for an organism. KEGG brite file will be download if needed.
}
\details{
Refer to \code{\link{kogs_table}} if you want gene to KOG mapping table.
}
\author{
ZG Zhao
}
