% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.map.R
\name{kogs_table}
\alias{kogs_table}
\title{Gene to KOG mapping table for an organism}
\usage{
kogs_table(org, d.path = "KEGG", KOGs = NA)
}
\arguments{
\item{org}{character, KEGG organism abbreviation}

\item{d.path}{file path for \code{\link{KEGG_get}}}

\item{KOGs}{character vector, names of KOGs. Subset if not empty.}
}
\value{
data.frame with three columns: gene, KOG and desc (functional description). NOTE: Some KEGG brite file may have no gene info, NULL is return without warning!
}
\description{
Get gene to KOG mapping table for an organism. KEGG brite file will be download if needed.
}
\details{
If org="ko", there are only two types of data: KOG (KEGG gene entries) and description. To provide a general interface, these data is also named "gene" and are duplicated in "ko" column. KEGG brite file for the organism will be downloaded automatically if not exists.
}
\author{
ZG Zhao
}
