% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.extra.R
\name{koInfo}
\alias{koInfo}
\title{Function koInfo}
\usage{
koInfo(kos = NULL, keyword = NULL, level = "B", children = FALSE)
}
\arguments{
\item{kos}{KO indices to filer the ko column of keggs object.}

\item{keyword}{keyword (regular expression) to filter the description column of keggs object.}

\item{level}{"A", "B", or "C". Used together with keyword.}

\item{children}{TRUE/FALSE, include children pathways or genes if TRUE.}
}
\value{
data.frame
}
\description{
Show or extract KEGG pathway info.
}
\details{
A data.frame object "keggs" is bound with the package. koInfo is used to filter this object.
}
\author{
ZG Zhao
}
