% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{make_mgraph}
\alias{make_mgraph}
\alias{make_bgraph}
\alias{make_rgraph}
\alias{make_ggraph}
\title{make graph from various data}
\usage{
- make_mgraph(object, ...)
- make_rgraph(object, ...)
- make_ggraph(object, ...)
}
\arguments{
\item{object}{keggPATH, ReactionSet ReactionList, character vector or list.}

\item{org}{character, organism/species indentifier, parameter for "object" without organism info: ReactionList and list}

\item{d.path}{character, refer to \code{\link{KEGG_get}} for detail. Used when "object" is a KO vecter (KOs).}
}
\description{
Make metabolic graph from various type of data: keggPATH, ReactionSet, ReactionList, list or KOs (KEGG pathway identifiers).
}
\details{
Metabolic network can be represented by metabolite, reaction or gene graph.
}
\author{
zhao
}
