% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpath.R
\name{make_mpath}
\alias{make_mpath}
\title{make metabolic pathway}
\usage{
make_mpath(object, d.path = "KEGG")
}
\arguments{
\item{object}{character, KEGG pathway ortholog (ko) index string such as ko00010, 00010 or ath00010}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
keggPATH object
}
\description{
Make metabolic pathway (keggPATH object)
}
\details{
This function reads from local folder (d.path) a KEGG xml file, which will be downloaded if not exists. Make sure internet is accessible.
Refer to \code{\link{keggPATH}} for class information.
TODO: make pathway from local KGML file regardless of d.path
}
\author{
ZG Zhao
}
