% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{make_rset}
\alias{make_rset}
\title{generate ReactionSet object}
\usage{
make_rset(object, org = "ko", d.path = "KEGG", ...)
}
\arguments{
\item{object}{KDataSet or list object. If feeding a list object, make sure each list elements are also lists containing substrate, product, gene and reversible elements.}

\item{org}{character, "ko" or organism identifier.}

\item{d.path}{file path for \code{\link{KEGG_get}}.}

\item{...}{other params}
}
\value{
ReactionSet object
}
\description{
Mostly for internal use.
}
\details{
"ReactionSet" is S4 class designed for mgraph. Reactions are presented in different forms from those of \code{\link{KDataSet}}.
}
\author{
ZG Zhao
}
