% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph_append}
\alias{mgraph_append}
\title{append reactions to mgraph manually}
\usage{
mgraph_append(g, df)
}
\arguments{
\item{g}{graphMET object}

\item{df}{a data frame with at least four columns which supply "from", "to", "reversible", "genes" infos for the reactions. File path to a csv file containing such infos is also acceptable.}
}
\value{

}
\description{
Append additional reactions to KEGG pathway.
}
\details{
TODO: may not work
Reactions without KEGG index cannot be extrated from KGML files. This function help you add these reactions manually.
}
\author{
ZG Zhao
}
