% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph_from_kmeta}
\alias{mgraph_from_kmeta}
\title{make graph from KEGGmeta object}
\usage{
mgraph_from_kmeta(kinfo)
}
\arguments{
\item{kinfo}{\code{\link{KEGGmeta}} object}
}
\value{
igraph object
}
\description{
Reconstruction of KEGG metabolic network represented by an mgraph object.
}
\details{
After conversion, compounds are represented by nodes and reactions by edges in the igraph object. Other information of KEGG metabolic networks are hold as attributes.
}
\author{
zhao
}
