% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph_from_kos}
\alias{mgraph_from_kos}
\title{make graph from kos}
\usage{
mgraph_from_kos(kos, d.path = "KEGG")
}
\arguments{
\item{kos}{character vector, KEGG pathway id(s). Mixing pathway ids of different organisms is not allowed.}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
mgraph object.
}
\description{
Reconstruct a KEGG pathway as mgraph network directly by providing the pathway identifiers (kos).
}
\details{
KGML file will be downloaded online if you do not supply.
}
\author{
ZG Zhao
}
