% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{mgraph_simplify}
\alias{mgraph_simplify}
\title{Simplify the structure of a metabolic network}
\usage{
mgraph_simplify(g)
}
\arguments{
\item{g}{metgraph object}
}
\value{
metgraph object
}
\description{
This function will: (1) merge any serial or parallel reactions; (2) update edge EP data accordingly.
}
\details{
A simplified network contains no serial or parallel reactions.
}
\author{
ZG Zhao
}
