% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mgraph.R
\name{mgraph_x_org}
\alias{mgraph_x_org}
\title{general to organism-specific mgraph conversion}
\usage{
mgraph_x_org(g, org, d.path = "KEGG")
}
\arguments{
\item{g}{mgraph object}

\item{org}{character}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
mgraph with organism set
}
\description{
Normalize KEGG generic pathway to species specific pathway.
}
\details{
Adjust and filter reactions, and rebuild graph object from reactions list.
}
\author{
ZG Zhao
}
