% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpath.R
\name{mpath_x_org}
\alias{mpath_x_org}
\title{General to organism-specific keggPATH conversion}
\usage{
mpath_x_org(kobj, org = NA, d.path = "KEGG")
}
\arguments{
\item{kobj}{a keggPATH object}

\item{org}{abbreviation of an organism such as "ath"}

\item{d.path}{file path for \code{\link{KEGG_get}}}
}
\value{
keggPATH object
}
\description{
Convert KOG names of a keggPATH object to real gene names
}
\details{
In keggPATH objects, genes involved in reactions are generally represented by KEGG entry names (k genes or KOG).
}
\author{
ZG Zhao
}
