% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{mreacts_append}
\alias{mreacts_append}
\title{append reaction}
\usage{
mreacts_append(robj, s, p, genes)
}
\arguments{
\item{robj}{MReactions object}

\item{s}{character, name of substrate}

\item{p}{character, name of product}

\item{genes}{character vector, names of genes involved in the reaction}
}
\value{
MReactions object
}
\description{
Append reaction to MReactions object
}
\author{
ZG Zhao
}
