% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotReaction}
\alias{plotReaction}
\title{show reaction diagram}
\usage{
plotReaction(
  substrates,
  products,
  r.name,
  shape.chem = "ellipse",
  shape.react = "rect",
  lcol = "gray",
  arr.pos = 0.7,
  arr.col = "gray",
  arr.tcol = "transparent",
  box.prop = 0.5,
  shadow.size = 0,
  relsize = 0.85,
  ...
)
}
\arguments{
\item{r.name}{reaction name}

\item{...}{pars passed to \code{\link{diagram::plotmat}}}

\item{substrate}{substrate names}

\item{product}{product names}
}
\description{
This function is not relevant to KEGG pathway. It is designed for illustration only.
}
\details{
Install "diagram" package manually before using this function.
}
\author{
zhao
}
