% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\name{prob_stcuts}
\alias{prob_stcuts}
\title{functional probability of s-t gene cutsets}
\usage{
prob_stcuts(g.sets, s.groups, GMR = 0.01)
}
\arguments{
\item{g.sets}{list, each item is a named gene vector and the names (set names) are used for grouping.}

\item{s.groups}{list, each item if a vector containg set names.}

\item{GMR}{numeric, gene mutation rate (default 0.01).}
}
\value{
numeric vector, functional probability of each gene group.
}
\description{
Calculate functional probability of a gene cutset
}
\details{
Note that a "functional" cutset of genes means all genes in it are mutated. If containing overlap genes, the cut sets must be grouped and joint probability is calculated.
}
\author{
ZG Zhao
}
