% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rcount}
\alias{rcount}
\alias{vcount}
\alias{ecount}
\title{count of graph entities}
\usage{
ecount(g)
vcount(g)
rcount(xg)
rcount(rset)
}
\arguments{
\item{g}{graph (igraph or xgraph) object}

\item{xg}{xgraph object}

\item{rset}{ReactionSet object}
}
\value{
character vector
}
\description{
Get count/number of nodes, edges or reactions
}
\details{
\itemize{
\item ecount: edges names (igraph)
\item vcount: vertex/node names (igraph)
\item rcount: number of unique reactions
}
}
\author{
ZG Zhao
}
