% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rdata}
\alias{rdata}
\title{reaction data}
\usage{
rdata(g, a.name, x.names)
}
\arguments{
\item{g}{mgraph or ReactionSet object}

\item{a.name}{character, attribute name}

\item{x.names}{vector of character (reaction names) or integer (ids)}
}
\description{
Get reaction data
}
\details{
Helper function for reaction data retrieving.
NOTE: please don't try to set reaction data alone; it may mess all thing up!
}
\seealso{
\code{\link{vdata}}
}
\author{
ZG Zhao
}
