% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rlist_append}
\alias{rlist_append}
\title{append reaction}
\usage{
rlist_append(object, x, p, genes, ids = NULL, reversible = FALSE)
}
\arguments{
\item{object}{ReactionList or equivalent list}

\item{x}{character (substrate name) or data.frame with at least these columns: from, to, genes, reversible}

\item{p}{character, product name (ignore if x is a data.frame)}

\item{genes}{character vector, names of genes involved in the reaction  (ignore if x is a data.frame)}

\item{ids}{additional identifers for reaction  (ignore if x is a data.frame)}

\item{reversible}{TRUE/FALSE (default). If TRUE, add reactions in both directions. Ignore if x is a data.frame.}
}
\value{
ReactionList object
}
\description{
Append reaction to reaction lists
}
\details{
Useful for adding to network spontaneous reactions without associated genes.
}
\author{
ZG Zhao
}
