% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rnames}
\alias{rnames}
\title{reaction names}
\usage{
rnames(object)
}
\arguments{
\item{object}{mgraph or MReactions object}
}
\description{
get reaction names
}
\details{
Similar functions: \code{\link{vnames}}, \code{\link{enames}}, \code{\link{rnames}}, \code{\link{vcount}}, \code{\link{ecount}}
}
\author{
ZG Zhao
}
