% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust}
\alias{robust}
\title{calculate mutational robustness}
\usage{
robust(object, GMR = 0.01, prob = FALSE)
}
\arguments{
\item{object}{mgraph with substrates and products set}

\item{GMR}{numeric (default 0.01), gene mutation rate/probability between 0 and 1.}

\item{prob}{TRUE/FALSE (default). Return functional probability of the network if TRUE.}
}
\value{
a number for robustness
}
\description{
Calculate mutational robustness of metabolic network
}
\details{
NOTE: you must set substrates and products before doing this!
}
\author{
ZG Zhao
}
