% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rset_append}
\alias{rset_append}
\title{append reaction}
\usage{
rset_append(robj, s, p, genes)
}
\arguments{
\item{robj}{RSet object}

\item{s}{character, name of substrate}

\item{p}{character, name of product}

\item{genes}{character vector, names of genes involved in the reaction}
}
\value{
RSet object
}
\description{
Append reaction to RSet object
}
\author{
ZG Zhao
}
