% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rset_merge_chems}
\alias{rset_merge_chems}
\title{merge chemicals in ReactionSet}
\usage{
rset_merge_chems(robj, chems, alias)
}
\arguments{
\item{robj}{ReactionSet object}

\item{chems}{vector, names of chemicals to merge}

\item{alias}{character, alias for chems after merge}
}
\value{
ReactionSet
}
\description{
DEPRECATED! Merge chemicals: rename chemicals and clean reactions.
}
\details{
Chemicals merge scheme:
\itemize{
\item rename "chems" to merge as "alias"
\item remove duplicated metabolites for each reaction
\item NOTE: remove reactions containg same metabolite in both substrates and products
\item rebuild ReactionSet (combine reactions?)
\item append alias information to new object
}
}
\author{
ZG Zhao
}
