% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{setorg_graph}
\alias{setorg_graph}
\title{KO-graph to species-specific graph}
\usage{
setorg_graph(object, org, d.path = "KEGG")
}
\arguments{
\item{object}{xgraph object}

\item{org}{character}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
graph with organism set
}
\description{
Normalize KEGG generic pathway to species specific pathway.
}
\details{
Adjust and filter reactions, and rebuild graph object from reactions list.
}
\examples{
## NOT RUN
library(syzMetab)
gm <- make_mgraph("ko00010")
gx <- subset_graph(gm, org="ath")
par(mfrow=c(2,1))
plot(gm)
## edges (reactions) with zero gene are show in dotted lines.
plot(gx)
}
\author{
ZG Zhao
}
