% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcuts.R
\name{setorg_stcuts}
\alias{setorg_stcuts}
\title{general (species="ko") stcuts to specices-specific stcuts}
\usage{
setorg_stcuts(object, org, d.path = "KEGG")
}
\arguments{
\item{object}{stcuts object}

\item{org}{character}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
stcuts
}
\description{
Convert gene names and filter empty sets
}
\details{
For time saving if you deal with a bunch of of organisms.
}
\author{
ZG Zhao
}
