% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{subset_graph}
\alias{subset_graph}
\title{subset xgraph}
\usage{
subset_graph(object, s, t, clean = FALSE)
}
\arguments{
\item{object}{graph object}

\item{s}{source nodes}

\item{t}{target nodes}
}
\value{
graph object
}
\description{
Subset graph with given sources (primary substrates) and targets (products)
}
\details{
\itemize{
\item In "setorg_graph", reactions without mapping genes are labeled with "absent". This function removes the "absent" links in the graph.
\item After setting sources and targets, nodes to sources or from targets should be removed.
\item Nodes not accessed by sources or targets should be removed also.
\item If clean=TRUE, nodes and edges not in paths from sources to targets are further removed by calling \code{\link{all_spaths_list}}.
}
}
\examples{
## NOT RUN
library(gmetab)
gm <- make_mgraph("ko00010")
chem1 <- c("C00031", "C00221", "C00267", "C01172", "C01451", "C06186")
chem2 <- "C00022"
gx <- subset_graph(gm, s=chem1, t=chem2)
par(mfrow=c(2,1))
plot(gm)
plot(gx)
}
\author{
ZG Zhao
}
