% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vdata}
\alias{vdata}
\title{node data}
\usage{
vdata(g, a.name, v.names)
}
\arguments{
\item{g}{igraph/mgraph object}

\item{a.name}{character, attribute name}

\item{v.names}{vector of character (node names) or integer (indices)}
}
\description{
Get or set node data/attributes
}
\details{
Friendly function for node data/attributes retrieving or setting. See "Example".
}
\examples{
library(gmetab)
d.path <- file.path(path.package("gmetab"), "KEGG")
gg <- make_mgraph("ko00010", d.path)
## igraph style
V(gg)$name
V(gg)$EP <- sample(vcount(gg))
V(gg)$EP[1:3]
## gmetab style
vdata(gg, "name")
vdata(gg, "name", 1:3)
(vv <- vnames(gg))
vdata(gg, "EP", "C00022")
vdata(gg, "EP") <- 1
vdata(gg, "EP")
}
\seealso{
\code{\link{edata}}
}
\author{
ZG Zhao
}
