% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vis_connected}
\alias{vis_connected}
\title{test connected}
\usage{
vis_connected(g, s, p)
}
\arguments{
\item{g}{graph object}

\item{s}{source node(s)}

\item{p}{target edge(s)}
}
\value{
logi, TRUE if any target is accessed by any source
}
\description{
test whether nodes are connected (have any route)
}
\details{
Note that non-connecting nodes are judged by infinite distance.
}
\author{
ZG Zhao
}
