% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vnames}
\alias{vnames}
\title{node names}
\usage{
vnames(object)
}
\arguments{
\item{object}{mgraph object}
}
\description{
get node names
}
\details{
Similar functions: \code{\link{vnames}}, \code{\link{enames}}, \code{\link{rnames}}, \code{\link{vcount}}, \code{\link{ecount}}
}
\author{
ZG Zhao
}
