% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vs_accessed_by}
\alias{vs_accessed_by}
\title{vertices assessed by given vertices}
\usage{
vs_accessed_by(g, v.names, mode = c("out", "in", "all"))
}
\arguments{
\item{g}{graph object}

\item{v.names}{character vector, names of vertices}

\item{mode}{character, "out" (default), "in" or "all", direction from given vertices.}
}
\value{
vertices vector
}
\description{
Get all vertices assessed by given vertices.
}
\details{
"Accessed" means that not all the distances from (mode="out") or to (mode="in") given vertices are infinite.
}
\author{
ZG Zhao
}
