% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vs_adjacent}
\alias{vs_adjacent}
\title{adjacent vertices}
\usage{
vs_adjacent(g, v.names, mode = c("out", "in", "all"))
}
\arguments{
\item{g}{graph object}

\item{v.names}{character vector, names of vertices}

\item{mode}{character, "out" (default), "in" or "all", direction from given vertices.}
}
\value{
named list of vertices
}
\description{
Get adjacent vertices
}
\details{
Refer to igraph::adjacent_vertices
}
\author{
ZG Zhao
}
