% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vsdelete}
\alias{vsdelete}
\alias{delete_vertices}
\alias{delete.vertices}
\title{delete vertices}
\usage{
vsdelete(object, vs)
}
\arguments{
\item{object}{igraph/mgraph object}

\item{vs}{vertices (indices or names)}
}
\value{
igraph/mgraph object
}
\description{
delete nodes/vertices from igraph or mgraph object
}
\details{
Unlike \code{\link{igraph::delete_vertices}}, graph attributes are retained after nodes removed.
}
\author{
ZG Zhao
}
