% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{xgraph_clean}
\alias{xgraph_clean}
\title{clean metabolic network}
\usage{
xgraph_clean(g, s, p, mc.cores = detectCores() - 1)
}
\arguments{
\item{g}{xgraph object}

\item{s}{character vector, primary substrate names}

\item{p}{character vecter, end product names}
}
\value{
mgraph object
}
\description{
Get "clean" metabolic network with given substrates and products.
}
\details{
Given primary substrates and end products, a "clean" metabolic network is a network in which all chemicals (vertices) and reactions (edges) are in the reaction chains from primary substrates to end products.
}
\author{
ZG Zhao
}
