% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{xgraph_orgset}
\alias{xgraph_orgset}
\title{KO-graph to species-specific graph}
\usage{
xgraph_orgset(object, org, d.path = "KEGG")
}
\arguments{
\item{object}{xgraph object}

\item{org}{character}

\item{d.path}{file path for \code{\link{KEGG_get}}.}
}
\value{
graph with organism set
}
\description{
Normalize KEGG generic pathway to species specific pathway.
}
\details{
Adjust and filter reactions, and rebuild graph object from reactions list.
}
\author{
ZG Zhao
}
