% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{xgraph_setchems}
\alias{xgraph_setchems}
\title{set substrates and products for a graph}
\usage{
xgraph_setchems(object, s, p, clean = FALSE)
}
\arguments{
\item{object}{xgraph object}

\item{s}{character vecter, names of primary substrates}

\item{p}{character vecter, names of end products}

\item{clean}{TRUE/FALSE(default). Remove futile nodes and edges if TRUE.}
}
\value{
xgraph object
}
\description{
Set chemicals and subset a xgraph (mgraph, rgraph or ggraph) object.
}
\details{
The function do these jobs:
\itemize{
\item for rgraph (reaction graph) and ggraph (gene graph), S and P are added to the graph as nodes.
\item all edges to S or from P are deleted, and then nodes not connecting to S or P are delete.
\item add "substrates" and "products" attributes are the graph.
}
}
\author{
ZG Zhao
}
