% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{xgraph_subset}
\alias{xgraph_subset}
\title{subset xgraph}
\usage{
xgraph_subset(object, s, t)
}
\arguments{
\item{object}{xgraph object}

\item{s}{character vecter, source ndoes}

\item{t}{character vecter, target nodes}
}
\value{
xgraph object
}
\description{
Subset graph with given sources (primary substrates) and targets (products)
}
\details{
The function do these jobs:
\itemize{
\item for rgraph (reaction graph) and ggraph (gene graph), S and P are added to the graph as nodes.
\item all edges to S or from P are deleted, and then nodes not connecting to S or P are delete.
\item add "substrates" and "products" attributes are the graph.
}
}
\author{
ZG Zhao
}
