% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_classes.R
\name{class_virtual}
\alias{class_virtual}
\alias{igraph}
\alias{mgraph}
\alias{KEntityList}
\title{virtual classes}
\description{
Virtual classes designed in mgraph package.
}
\details{
There following classes are set as virutal:
\itemize{
\item igraph
\item mgraph
\item KEntityList: actually list
}
}
\author{
zhao
}
