% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{delete.vertices}
\alias{delete.vertices}
\alias{delete_vertices}
\alias{add.vertices}
\alias{add_vertices}
\title{delete vertices}
\usage{
delete.vertices(object, vs)
}
\value{
igraph/mgraph object
}
\description{
delete vertices/nodes from igraph or mgraph object
}
\details{
Same as functions in igraph package exception for retaining graph attributes. Refer to \code{?igraph::delete.vertices}, \code{?igraph::delete.vertices} for details.
}
\author{
ZG Zhao
}
