% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{helpers_general}
\alias{helpers_general}
\alias{getOrganism}
\alias{getCPDs}
\alias{getGenes}
\alias{getReactions}
\alias{getAliases}
\alias{Reactions}
\alias{Compounds}
\alias{Species}
\alias{Organism}
\alias{Genes}
\alias{Aliases}
\title{general helper function}
\description{
general helpers for retrieving data from keggPATH, reactions list or mgraph object
}
\details{
All functions below are generic functions. Function names are self-explanatory:
\itemize{
\item pathInfo(object)
\item getOrganism(object)
\item getGenes(object): all genes/orthologs
\item getCPDs(object): all compounds
\item getReactions(object, list.only=TRUE): reactions list or ReactionSet object (list.only=FALSE for mgraph object)
\item getAliases(object)
}

Alias functions:
\itemize{
\item Organism(object)
\item Species(object)
\item Compounds(object)
\item Chemicals(object)
\item Genes(object)
\item Reactions(object, list.only=TRUE)
\item Aliases(object)
}
}
\examples{
library(gmetab)
d.path <- file.path(path.package("gmetab"), "KEGG")
kinfo <- make_mpath("ko00010", d.path)
gg <- mgraph_from_mpath(kinfo)
##
pathInfo(kinfo)
getOrganism(kinfo)
Organism(kinfo)
getCPDs(kinfo)
Compounds(kinfo)
Chemicals(kinfo)
getGenes(kinfo)
# str(getReactions(kinfo))
# str(Reactions(kinfo))
##
getOrganism(gg)
Organism(gg)
getCPDs(gg)
Compounds(gg)
Chemicals(gg)
Genes(gg)
# getReactions(gg)
# Reactions(gg)
Reactions(gg, list.only=FALSE)
getAliases(gg)
Aliases(gg)
}
\author{
ZG Zhao
}
