% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{make_mgraph}
\alias{make_mgraph}
\alias{class_mgraph}
\title{make graph from keggPATH object}
\usage{
make_mgraph(object, org=NULL)
}
\arguments{
\item{object}{keggPATH, ReactionSet ReactionList, list or even igrph object.}

\item{org}{character, organism/species indentifier, parameter for "object" without organism info: ReactionList and list}

\item{d.path}{character, refer to \code{\link{KEGG_get}} for detail. Used when "object" is a KO vecter (KOs).}
}
\description{
Make mgraph object from various type of data: keggPATH, ReactionSet, ReactionList, list or KOs (KEGG pathway identifiers).
}
\details{
"mgraph" is a virtual S3 class defined in "gmetab" package.
\itemize{
\item A mgraph object is actually a igraph object with a key attribute: "reactions", which is vital for downstream metwork analysis.
\item Edges of a mgraph object have a "reactions" attribute holding the names of reactions involved, helpful for mapping edges back to reactions and genes.
\item Users can set other graph, node or edge attributes
}
}
\author{
zhao
}
