% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph.R
\name{make_mgraph}
\alias{make_mgraph}
\alias{make_bgraph}
\alias{make_rgraph}
\alias{make_ggraph}
\title{make graph from various data}
\usage{
- make_mgraph(object, ...): from KDataSet, ResctionSet, ReactionList, list or character vector (kos)
- make_rgraph(object, ...): from KDataSet, ResctionSet, ReactionList or character vector (kos)
- make_ggraph(object, ...): from KDataSet, ResctionSet, ReactionList or character vector (kos)
- make_bgraph(object, ...): from mgraph or stcuts (refer to \code{\link{make_stcuts}})
}
\arguments{
\item{object}{KDataSet, ReactionSet ReactionList, character vector or list.}

\item{...}{additional parameters depends on "object"
\itemize{
\item org: organism/species indentifier, parameter for "object" without organism info such as ReactionList and list
\item d.path: character, refer to \code{\link{KEGG_get}} for detail. Used when "object" is a character vector (kos).
}}
}
\description{
Make metabolic graph from various type of data: KDataSet, ReactionSet, ReactionList, list, KOs or other xgraph object.
}
\details{
Metabolic network can be represented by metabolite, reaction, gene graph or st-cuts of gene sets.
\itemize{
\item make_mgraph (make metabolite graph): nodes are metabolites and edges are reactions identified by gene sets
\item make_rgraph (make reaction graph): nodes are reactions
\item make_ggraph (make gene graph): nodes are genes
\item make_bgraph (make bipartite graph): derived from st-cuts of mgraph. Nodes are genes and st-cuts of gene sets. Edges are links between gene sets and the genes included.
}
}
\examples{
## NOT RUN
library(syzMetab)
pp <- make_kdset("ko00010", d.path="KEGG")
make_mgraph(pp)
make_mgraph(Reactions(pp), org="ko")
make_mgraph("ko00010")
gm <- make_mgraph(c("ko00010", "ko00020"))
plot(gm)
gr <- make_rgraph("ko00010")
plot(gr)
gg <- make_ggraph("ko00010")
plot(gg)
chem1 <- c("C00031", "C00221", "C00267", "C01172", "C01451", "C06186")
chem2 <- "C00022"
gb <- make_bgraph(gm, s=chem1, t=chem2)
plot(gb)
}
\author{
zhao
}
