% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcuts.R
\name{make_stcuts}
\alias{make_stcuts}
\title{list all gene cut sets}
\usage{
make_stcuts(object, s, t)
}
\arguments{
\item{object}{xgraph object}

\item{s}{source node}

\item{t}{target node}
}
\value{
list
\itemize{
\item min.genes
\item min.edges
\item edge.cuts
\item gene.cuts
}
}
\description{
Get st_cuts info: edges and genes
}
\details{
Get st-cuts of edges (reactions). The edge cut-sets are translated into gene cut-sets also. Results can apply to \code{\link{make_bgraph}}.
}
\examples{
## NOT RUN
library(gmetab)
gm <- make_mgraph(c("ko00010", "ko00020"))
plot(gm)
chem1 <- c("C00031", "C00221", "C00267", "C01172", "C01451", "C06186")
chem2 <- "C00022"
gb <- make_bgraph(gm, s=chem1, t=chem2)
plot(gb)
xcuts <- make_stcuts(gm, s=chem1, t=chem2)
str(xcuts)
gx <- make_bgraph(xcuts)
plot(gx)
}
\author{
ZG Zhao
}
