% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mgraph}
\alias{plot.mgraph}
\alias{plot.ggraph}
\title{plot mgraph}
\usage{
\method{plot}{mgraph}(g, show.name = TRUE, vlcex = 1, ...)
}
\arguments{
\item{g}{mgraph or igraph object}

\item{show.name}{logi. Show chemical names if TRUE (default).}

\item{...}{pars passed to plot.igraph}
}
\description{
plot mgraph
}
\details{
Specified igraph plot function for mgraph
}
\author{
zhao
}
