% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mgraph}
\alias{plot.mgraph}
\alias{plot.rgraph}
\alias{plot.ggraph}
\alias{plot.bgraph}
\title{plot graph}
\usage{
\method{plot}{mgraph}(g, s, t, p = t, show.name = TRUE, gene.n = FALSE, ...)
}
\arguments{
\item{g}{xgraph object}

\item{s}{substrates}

\item{t}{targets, same as "p"}

\item{p}{products}

\item{show.name}{logi. Show chemical names if TRUE (default).}

\item{gene.n}{TRUE/FALSE (default). show gene number on edge if TRUE.}

\item{...}{pars passed to plot.igraph}
}
\description{
plot graph
}
\details{
Specified igraph plot function for mgraph
}
\author{
zhao
}
