% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rcount}
\alias{rcount}
\title{reaction names}
\usage{
rcount(object)
}
\arguments{
\item{object}{mgraph or ReactionSet object}
}
\value{
integer, the number of unique reactions listed in "ReactionList" of the object.
}
\description{
get number of reactions
}
\details{
Similar to vcount and ecount in igraph package.
}
\author{
ZG Zhao
}
