% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rdata}
\alias{rdata}
\title{reaction data}
\usage{
rdata(g, a.name, x.names)
}
\arguments{
\item{g}{mgraph or RSet object}

\item{a.name}{character, attribute name}

\item{x.names}{vector of character (reaction names) or integer (ids)}
}
\description{
Get or set reaction data
}
\details{
Friendly function for edge data/attributes retrieving or setting. See "Example".
}
\seealso{
\code{\link{vdata}}
}
\author{
ZG Zhao
}
