% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rdata}
\alias{rdata}
\alias{edata}
\alias{vdata}
\title{get/set attribute data}
\usage{
vdata(g, a.name, v.name)
edata(g, a.name, e.name)
rdata(g, a.name, r.name)
vdata(g, a.name) <- vals
edata(g, a.name) <- vals
rdata(g, a.name) <- vals
}
\arguments{
\item{g}{igraph/mgraph object}

\item{a.name}{attribute name}

\item{v.names}{(optional) node names or indices}

\item{e.names}{(optional) edge names or indices}

\item{r.names}{(optional) reaction names (for subseting)}
}
\description{
Helper functions attribute data
}
\details{
Provide alternative ways to get or set graph attributes.
}
\author{
ZG Zhao
}
