% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rnames}
\alias{rnames}
\title{reaction names}
\usage{
rnames(object)
}
\arguments{
\item{object}{mgraph or ReactionSet object}
}
\value{
character vector. For \code{ReactionSet}, it is names of the reactions; for \code{xgraph} object, it is the reaction names of the edges, extracted from "reaction" edge attribute.
}
\description{
get reaction names
}
\details{
Similar functions: \code{\link{vnames}}, \code{\link{enames}}, \code{\link{rnames}}, \code{\link{vcount}}, \code{\link{ecount}}
}
\author{
ZG Zhao
}
