% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rnames}
\alias{rnames}
\alias{vnames}
\alias{enames}
\title{names of graph entities}
\usage{
enames(g)
vnames(g)
rnames(xg)
rnames(rset)
}
\arguments{
\item{g}{graph (igraph or xgraph) object}

\item{xg}{xgraph object}

\item{rset}{ReactionSet object}
}
\value{
character vector
}
\description{
Get names (instead of ids) of nodes, edges or reactions (xgraph only).
}
\details{
\itemize{
\item enames: edges names (igraph)
\item vnames: vertex/node names (igraph)
\item rnames: For graph object, it returns the reaction names (may be duplicated) along vnames. For "ReactionSet" object, they are the (unique) names of reactions.
}
}
\author{
ZG Zhao
}
