% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{rset_append}
\alias{rset_append}
\title{append reaction}
\usage{
rset_append(robj, s, p, genes, ids = NULL)
}
\arguments{
\item{robj}{ReactionSet object}

\item{s}{character, name of substrate}

\item{p}{character, name of product}

\item{genes}{character vector, names of genes involved in the reaction}

\item{ids}{additional identifers for reaction}
}
\value{
ReactionSet object
}
\description{
Append reaction to ReactionSet object
}
\details{
Useful for adding to network spontaneous reactions without associated genes.
}
\author{
ZG Zhao
}
