% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpath.R
\name{setorg_path}
\alias{setorg_path}
\title{General to organism-specific KDataSet conversion}
\usage{
setorg_path(object, org, d.path = "KEGG")
}
\arguments{
\item{object}{a KDataSet object}

\item{org}{abbreviation of an organism such as "ath"}

\item{d.path}{file path for \code{\link{KEGG_get}}}
}
\value{
KDataSet object
}
\description{
Convert KOG names of a KDataSet object to real gene names
}
\details{
In KDataSet objects, genes involved in reactions are generally represented by KEGG entry names (k genes or KOG).
}
\author{
ZG Zhao
}
