% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{vsdelete}
\alias{vsdelete}
\alias{delete_vertices}
\alias{delete.vertices}
\title{delete vertices}
\usage{
vsdelete(object, vs)
}
\arguments{
\item{object}{igraph/mgraph object}

\item{vs}{vector: vertex ids (integer) or names (character)}
}
\value{
igraph/mgraph object
}
\description{
delete vertices/nodes from igraph or mgraph object
}
\details{
Refer to \code{\link{igraph::delete.vertices}}
}
\author{
ZG Zhao
}
