% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgraph.R
\name{xgraph_setchems}
\alias{xgraph_setchems}
\title{set substrates and products for a graph}
\usage{
xgraph_setchems(object, s, p, mc.cores, n.max)
}
\arguments{
\item{object}{xgraph object}

\item{s}{character vecter, names of primary substrates}

\item{p}{character vecter, names of end products}

\item{mc.cores}{integer, for \code{\link{all_spath_list}}, refer to \code{\link{mclapply}} for details.}

\item{n.max}{numeric, for \code{\link{all_spath_list}}.}
}
\value{
xgraph object
}
\description{
Set chemicals and subset a xgraph (mgraph, rgraph or ggraph) object.
}
\details{
The function do these jobs:
\itemize{
\item for rgraph (reaction graph) and ggraph (gene graph), S and P are added to the graph as nodes.
\item all metabolites (nodes) having no contribution in forming of P from S are deleted, and so do the reactions (edges).
This is done by screening nodes which are in at least one simple pathway from S to P.
\item add "substrates" and "products" attributes are the graph.
}
}
\author{
ZG Zhao
}
